Add files via upload
I ensure that my code, program or software is functional and secure by making these repositories private or encrypting them to specific servers if not on github. I will continue to do automatic testing to increase efficiency and save time. Interpreting user needs and incorporating them into a program as easily done by determining what is the missing feature and adding it as a benefit in the code. I approach the software by blueprinting what is needed and making sure it hits all of the requirements.